+++ ParticlesInABox/collisions.h
@@ -2,6 +2,6 @@
 
 #include "core.h"
 
-float circularBodyTimeInstance(object* ball1, object* ball2);
+double circularBodyTimeInstance(object* ball1, object* ball2);
 
-float trajectory(float x, float xp, float t);
+double trajectory(double x, double xp, double t);
